[DOCS] Format TypoScript reference differently
The table is now formatted with the directive confval which is
a supported directive and is more recently used for references
in the TYPO3 documentation.

The previous table had rendering problems and was not visible
on the rendered page.

Additionally, tables have the problem that they are tedious to
create and maintain and are often not rendered well on small
screens - either cropped or rendered with scrollbars.

Resolves: #8